🤾🏼‍♂️ Merge entry and exit utilities into one
