--- entry/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "entry",
-  "version": "1.0.0",
-  "description": "Entry point that relays Ethereum RPC messages through the Nym network.",
-  "main": "index.js",
-  "author": "Daniel Luca (CleanUnicorn)",
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@types/node": "^18.11.18",
-    "express": "^4.18.2",
-    "typescript": "^4.9.4",
-    "ws": "^8.11.0"
-  },
-  "scripts": {
-    "build": "tsc",
-    "start": "npm run build && node dist/index.js",
-    "lint": "eslint . --ext .ts",
-    "lint:fix": "eslint . --ext .ts --fix",
-    "prettier-format": "prettier --config .prettierrc 'src/**/*.ts' --write"
-  },
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.48.0",
-    "@typescript-eslint/parser": "^5.48.0",
-    "eslint": "^8.31.0",
-    "prettier": "^2.8.1"
-  }
-}
