--- exit/.prettierrc
@@ -1,7 +1,0 @@
-{
-    "semi": true,
-    "trailingComma": "all",
-    "singleQuote": true,
-    "printWidth": 120,
-    "tabWidth": 4,
-  }
