--- exit/src/interfaces/Response.ts
@@ -1,7 +1,0 @@
-interface Response {
-    type: string;
-    message: string;
-    address: string;
-}
-
-export { Response };
