+++ src/exit/index.ts
