+++ src/exit/log/index.ts
